# backend/internal/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  pointmarket_backend_internal_dtos.APIError:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  pointmarket_backend_internal_dtos.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  pointmarket_backend_internal_dtos.LikertResultDTO:
    properties:
      answers:
        items:
          type: integer
        type: array
      created_at:
        type: string
      id:
        type: integer
      questionnaire_id:
        type: integer
      student_id:
        type: string
      subscale_scores:
        additionalProperties:
          type: number
        type: object
      total_score:
        type: number
    type: object
...
  pointmarket_backend_internal_dtos.MultimodalThresholdRequest:
    properties:
      threshold:
        type: number
    required:
    - threshold
    type: object
  pointmarket_backend_internal_dtos.MultimodalThresholdResponse:
    properties:
      threshold:
        type: number
    type: object
  pointmarket_backend_internal_dtos.NullData:
    type: object
...
  pointmarket_backend_internal_dtos.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/pointmarket_backend_internal_dtos.PaginationMeta'
      success:
        type: boolean
    type: object
  pointmarket_backend_internal_dtos.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
...
  pointmarket_backend_internal_dtos.ProgramDTO:
    properties:
      created_at:
        type: string
      faculty_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  pointmarket_backend_internal_dtos.StudentDTO:
    properties:
      birth_date:
        type: string
      cohort_year:
        type: integer
      created_at:
        type: string
      gender:
        type: string
      phone:
        type: string
      program:
        $ref: '#/definitions/pointmarket_backend_internal_dtos.ProgramDTO'
      status:
        type: string
      student_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  pointmarket_backend_internal_dtos.StudentDetailsDTO:
    properties:
      academic_score:
        type: number
      ams_result:
        $ref: '#/definitions/pointmarket_backend_internal_dtos.LikertResultDTO'
      birth_date:
        type: string
      cohort_year:
        type: integer
      display_name:
        type: string
      email:
        type: string
      gender:
        type: string
      mslq_result:
        $ref: '#/definitions/pointmarket_backend_internal_dtos.LikertResultDTO'
      phone:
        type: string
      program_name:
        type: string
      status:
        type: string
      student_id:
        type: string
      user_id:
        type: integer
      vark_result:
        $ref: '#/definitions/pointmarket_backend_internal_dtos.StudentLearningStyle'
    type: object
  pointmarket_backend_internal_dtos.StudentLearningStyle:
    properties:
      label:
        type: string
      scores:
        $ref: '#/definitions/pointmarket_backend_internal_dtos.VARKScores'
      type:
        type: string
    type: object
  pointmarket_backend_internal_dtos.StudentListItem:
    properties:
      cohort_year:
        type: integer
      created_at:
        type: string
      email:
        type: string
      name:
        type: string
      program:
        $ref: '#/definitions/pointmarket_backend_internal_dtos.ProgramDTO'
      status:
        type: string
      student_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  pointmarket_backend_internal_dtos.UpsertStudentRequest:
    properties:
      birth_date:
        type: string
      cohort_year:
        type: integer
      gender:
        type: string
      phone:
        type: string
      program_id:
        type: integer
      status:
        type: string
      student_id:
        type: string
    required:
    - program_id
    - student_id
    type: object
...
  pointmarket_backend_internal_dtos.VARKScores:
    properties:
      auditory:
        type: number
      kinesthetic:
        type: number
      reading:
        type: number
      visual:
        type: number
    type: object
...
info:
  contact: {}
  description: PointMarket backend API documentation.
  title: PointMarket API
  version: "1.0"
paths:
...
  /programs:
    get:
      description: Lists all academic programs (auth required)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/pointmarket_backend_internal_dtos.ProgramDTO'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: List academic programs
      tags:
      - programs
...
  /scorings/multimodal:
    get:
      description: Returns current multimodal threshold.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.MultimodalThresholdResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Get multimodal threshold
      tags:
      - scorings
    put:
      consumes:
      - application/json
      description: Updates current multimodal threshold.
      parameters:
      - description: Threshold
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pointmarket_backend_internal_dtos.MultimodalThresholdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.NullData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Update multimodal threshold
      tags:
      - scorings
  /students:
    get:
      description: Lists students with filters and pagination (admin-only)
      parameters:
      - description: Search query
        in: query
        name: search
        type: string
      - description: Student status
        in: query
        name: status
        type: string
      - description: Program ID
        in: query
        name: program_id
        type: integer
      - description: Cohort year
        in: query
        name: cohort_year
        type: integer
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/pointmarket_backend_internal_dtos.StudentListItem'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Search students
      tags:
      - students
  /students/{user_id}:
    get:
      description: Fetches a student record by user ID (admin-only)
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.StudentDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Get student by user ID
      tags:
      - students
    put:
      consumes:
      - application/json
      description: Creates or updates a student record for the given user ID (admin-only)
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Student payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pointmarket_backend_internal_dtos.UpsertStudentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Create or update student
      tags:
      - students
...
  /students/{user_id}/details:
    get:
      description: Fetches detailed student info, including latest questionnaire results
        (admin-only)
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.StudentDetailsDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Get student details
      tags:
      - students
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"